뷰와 카테고리 목록조회 api 연동
- StateNotifier 도입
- SliverList 사용

+++ lib/model/cmu/common/scroll_response.dart
@@ -1,8 +1,7 @@
-import 'dart:ffi';
 
 class ScrollResponse<T> {
   final List<T> feeds;
-  final Long lastCursorId;
+  final int lastCursorId;
   final bool hasNext;
 
   ScrollResponse({required this.feeds, required this.lastCursorId, required this.hasNext});
